Resolving comments related updates
